Add files via upload
If I were ever to publish the full mod, that's the name I picked. However, this file includes the Generators Operate on Consumable Fuel mod set. Map text might be unnecessary, or it is possible that I was trying to add a space station and a storyline tied to it so I recommend that you take a look. Every ship in the game was given a Generator Installment Allowance so that you may use only single Generator at a time, however in the power.txt, only Fusion Reactor requires it, probably because I had written the other reactors with more than 2 capacity modifiers and an outfit can't have a 3rd one. If I continued the project, I would simply remove certain capacity modifiers like that of Nuclear Fission Reactor MK-I's coolant slots to make space for the GIA. Missions.txt has a couple of dialogs for the space station I was planning to add. Dialogs ought to be sloppy, since I'm not really good at that sorf of thing.